fix remaining lint errors
