+++ package.json
@@ -37,7 +37,7 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.3.1",
-    "@relaycorp/eslint-config": "^1.1.48",
+    "@relaycorp/eslint-config": "^1.1.50",
     "@relaycorp/shared-config": "^1.9.2",
     "@types/jest": "^29.2.4",
     "@typescript-eslint/parser": "^5.46.1",
