+++ src/lib/oids.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/naming-convention */
 /**
  * ASN.1 Object Ids.
  */
