+++ src/lib/utils/keys/RsaPssPrivateKey.spec.ts
@@ -1,0 +1,32 @@
+import { type HashingAlgorithm } from '../algorithms.js';
+import { AwalaAesKwProvider } from '../webcrypto/AwalaAesKwProvider.js';
+import { MockAesKwProvider } from '../../../testUtils/webcrypto/MockAesKwProvider.js';
+
+import { RsaPssPrivateKey } from './RsaPssPrivateKey.js';
+
+const PROVIDER = new AwalaAesKwProvider(new MockAesKwProvider());
+
+const HASHING_ALGORITHM: HashingAlgorithm = 'SHA-384';
+
+describe('RsaPssPrivateKey', () => {
+  test('Key usages should only allow signing', () => {
+    const key = new RsaPssPrivateKey(HASHING_ALGORITHM, PROVIDER);
+
+    expect(key.usages).toStrictEqual(['sign']);
+  });
+
+  test('Hashing algorithm should be added to key algorithm', () => {
+    const key = new RsaPssPrivateKey(HASHING_ALGORITHM, PROVIDER);
+
+    expect(key.algorithm).toStrictEqual({
+      hash: { name: HASHING_ALGORITHM },
+      name: 'RSA-PSS',
+    });
+  });
+
+  test('Provider should be honoured', () => {
+    const key = new RsaPssPrivateKey(HASHING_ALGORITHM, PROVIDER);
+
+    expect(key.provider).toStrictEqual(PROVIDER);
+  });
+});
