+++ src/lib/utils/webcrypto/engine.ts
@@ -1,7 +1,7 @@
 import { CryptoEngine } from 'pkijs';
 import { type ProviderCrypto } from 'webcrypto-core';
 
-import { PrivateKey } from '../PrivateKey.js';
+import { PrivateKey } from '../keys/PrivateKey.js';
 
 import { VeraCrypto } from './VeraCrypto.js';
