+++ src/lib/utils/x509/Certificate.ts
@@ -17,7 +17,7 @@ import { derDeserialize } from '../asn1.js';
 import { generateRandom64BitValue } from '../crypto.js';
 
 import CertificateError from './CertificateError.js';
-import type FullCertificateIssuanceOptions from './FullCertificateIssuanceOptions.js';
+import type CertificateIssuanceOptions from './CertificateIssuanceOptions.js';
 
 const X509_CERTIFICATE_VERSION_3 = 2;
 
@@ -125,7 +125,7 @@ export default class Certificate {
   /**
    * Issue a Relaynet PKI certificate.
    */
-  public static async issue(options: FullCertificateIssuanceOptions): Promise<Certificate> {
+  public static async issue(options: CertificateIssuanceOptions): Promise<Certificate> {
     // PKI.js should round down to the nearest second per X.509. We should do it ourselves to
     // avoid discrepancies when the validity dates of a freshly-issued certificate are used.
     const validityStartDate = setMilliseconds(options.validityStartDate ?? new Date(), 0);
