+++ src/lib/utils/x509/CertificateIssuanceOptions.ts
@@ -1,7 +1,7 @@
 import type BasicCertificateIssuanceOptions from './BasicCertificateIssuanceOptions.js';
 import type Certificate from './Certificate.js';
 
-export default interface FullCertificateIssuanceOptions extends BasicCertificateIssuanceOptions {
+export default interface CertificateIssuanceOptions extends BasicCertificateIssuanceOptions {
   readonly isCa?: boolean; // Basic Constraints extension
   readonly commonName: string;
   readonly issuerCertificate?: Certificate; // Absent when self-signed
