+++ src/testUtils/pki.ts
@@ -1,15 +1,15 @@
 import Certificate from '../lib/utils/x509/Certificate.js';
 import { generateRsaKeyPair, getPublicKeyDigestHex } from '../lib/utils/keys.js';
-import type FullCertificateIssuanceOptions from '../lib/utils/x509/FullCertificateIssuanceOptions.js';
+import type CertificateIssuanceOptions from '../lib/utils/x509/CertificateIssuanceOptions.js';
 
 interface StubCertConfig {
-  readonly attributes: Partial<FullCertificateIssuanceOptions>;
+  readonly attributes: Partial<CertificateIssuanceOptions>;
   readonly issuerCertificate: Certificate;
   readonly issuerPrivateKey: CryptoKey;
   readonly subjectPublicKey: CryptoKey;
 }
 
-export function reSerializeCertificate(cert: Certificate): Certificate {
+function reSerializeCertificate(cert: Certificate): Certificate {
   // TODO: Raise bug in PKI.js project
   // PKI.js sometimes tries to use attributes that are only set *after* the certificate has been
   // deserialized, so you'd get a TypeError if you use a certificate you just created in memory.
@@ -26,11 +26,11 @@ export async function generateStubCert(config: Partial<StubCertConfig> = {}): Pr
   const futureDate = new Date();
   futureDate.setDate(futureDate.getDate() + 1);
   futureDate.setMilliseconds(0);
-  const subjectPublicKey = config.subjectPublicKey || keyPair.publicKey;
+  const subjectPublicKey = config.subjectPublicKey ?? keyPair.publicKey;
   const certificate = await Certificate.issue({
     commonName: `0${await getPublicKeyDigestHex(subjectPublicKey)}`,
     issuerCertificate: config.issuerCertificate,
-    issuerPrivateKey: config.issuerPrivateKey || keyPair.privateKey,
+    issuerPrivateKey: config.issuerPrivateKey ?? keyPair.privateKey,
     subjectPublicKey,
     validityEndDate: futureDate,
     ...config.attributes,
